Extract the setting DSL into ConfigDsl
Config keeps the declarations and the imgproxy check; the generic
machinery moves to a module. Declaring the same name twice now
raises instead of silently overwriting.
