+++ config/application.rb
@@ -36,7 +36,7 @@ module F2Rails
     # Please, add to the `ignore` list any other `lib` subdirectories that do
     # not contain `.rb` files, or that should not be reloaded or eager loaded.
     # Common ones are `templates`, `generators`, or `middleware`, for example.
-    config.autoload_lib(ignore: %w[assets tasks config.rb])
+    config.autoload_lib(ignore: %w[assets tasks config.rb config_dsl.rb])
 
     # Configuration for the application, engines, and railties goes here.
     #
