+++ lib/config.rb
@@ -1,12 +1,13 @@
 require "uri"
+require_relative "config_dsl"
 
 # Registry for external configuration. Every credential or environment
 # setting the app consumes is declared here exactly once and read only through
-# this class (Config.resend_api_key, Config.imgproxy?).
+# this class (Config.resend_api_key, Config.imgproxy?). The declaration
+# machinery lives in ConfigDsl.
 #
-# Each setting resolves as source, then default, then normalize, then
-# validate. required and validate are independent: an optional setting that is
-# set still gets validated, because misconfigured is not the same as missing.
+# required and validate are independent: an optional setting that is set
+# still gets validated, because misconfigured is not the same as missing.
 # normalize canonicalizes a present value, so consumers never see raw quirks
 # such as trailing slashes.
 #
@@ -15,16 +16,15 @@ require "uri"
 # initialization, so a misconfigured process fails to start. Checks are local
 # only. Violation messages never include setting values.
 class Config
-  class ConfigurationError < StandardError; end
+  extend ConfigDsl
 
-  Setting = Data.define(:name, :source, :required, :default, :normalize, :validate)
+  class ConfigurationError < StandardError; end
 
   IMGPROXY_SETTINGS = %i[imgproxy_endpoint imgproxy_key imgproxy_salt].freeze
 
   class << self
     def validate!
-      violations = settings.each_value.flat_map { |setting| setting_violations(setting) }
-      violations += imgproxy_violations
+      violations = settings_violations + imgproxy_violations
       return if violations.none?
 
       raise ConfigurationError, "Invalid application configuration:\n#{violations.map { |text| "- #{text}" }.join("\n")}"
@@ -34,50 +34,7 @@ class Config
       IMGPROXY_SETTINGS.all? { |name| public_send("#{name}?") }
     end
 
-    # Effective configuration for the system status page. Booleans only,
-    # never values: most settings are secrets.
-    def status
-      settings.keys.index_with { |name| public_send("#{name}?") }
-    end
-
     private
-
-    def settings
-      @settings ||= {}
-    end
-
-    def setting(name, source:, required: false, default: nil, normalize: nil, validate: nil)
-      settings[name] = Setting.new(name:, source:, required:, default:, normalize:, validate:)
-
-      define_singleton_method(name) { value_of(name) }
-      define_singleton_method("#{name}?") { public_send(name).present? }
-    end
-
-    def value_of(name)
-      setting = settings.fetch(name)
-      value = setting.source.call
-      value = resolve(setting.default) if value.blank?
-      value = setting.normalize.call(value) if setting.normalize && value.present?
-      value
-    end
-
-    def resolve(value)
-      value.respond_to?(:call) ? value.call : value
-    end
-
-    def setting_violations(setting)
-      value = value_of(setting.name)
-
-      if value.blank?
-        return resolve(setting.required) ? ["#{setting.name}: required in #{Rails.env} but not set"] : []
-      end
-
-      return [] if setting.validate.nil? || setting.validate.call(value)
-
-      ["#{setting.name}: present but invalid"]
-    rescue StandardError => e
-      ["#{setting.name}: evaluation raised #{e.class}"]
-    end
 
     # imgproxy is configured all-or-nothing: off when every setting is absent,
     # on when all are present, and anything in between fails the gate. A
